refactor : MovieDataInitializer 수정, ShowtimeDataInitializer 에 모든 데이터 넣도록 수정

+++ src/main/java/com/example/letmovie/domain/movie/controller/ReviewController.java
@@ -18,7 +18,7 @@ public class ReviewController {
     private final ReviewServiceImpl reviewService;
 
     @PostMapping("/add")
-    public String addReview(@RequestParam int movieId,
+    public String addReview(@RequestParam Long movieId,
                             @RequestParam String nickname,
                             @RequestParam String password,
                             @RequestParam int rating,
